Use longer sample for random VAD test
The voice activity detector will not do much with short input samples
because it will be too short for various thresholds. Extend the input
tensor (for the random test) to ensure any latent issues are detected.

This now causes the VAD consistency test to fail when `batch_size=3`, so
it does seem to have identified a new issue.